Use unsigned long int to void* conversion.
git-svn-id: svn://svn.code.sf.net/p/symbol-lookup/code/trunk@223 43bcf1f2-bc33-0410-b139-a31c8fc47933

+++ portageutils.c
@@ -113,7 +113,7 @@ void find_ebuilds(const struct str_t *const file)
 
     /* hash found files */
     ENTRY entry;
-    for (unsigned int i=0; i<file->size; i++)
+    for (unsigned long int i=0; i<file->size; i++)
     {
         entry.key = file->str[i];
         entry.data = (void*)i;
